[Refactor] Change evaluation script path
