+++ docs/api-reference/app/chat.mdx
