+++ docs/api-reference/app/delete.mdx
