+++ docs/api-reference/app/reset.mdx
