+++ docs/api-reference/app/search.mdx
